requirements.txt
weaviate-client==3.26.0
google-generativeai==0.8.3
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.5
python-dotenv==1.0.1
matplotlib==3.9.2
seaborn==0.13.2
datasets==2.19.0

